refactor : dto안에 있던 enum값 유효성 메서드를 제거후, Facility 생성자 안에서 메서드 호출하도록 변경
