+++ src/main/java/com/example/rentalSystem/domain/facility/dto/request/CreateFacilityRequestDto.java
@@ -1,9 +1,6 @@
 package com.example.rentalSystem.domain.facility.dto.request;
 
-import com.example.rentalSystem.domain.facility.entity.Facility;
-import com.example.rentalSystem.domain.facility.entity.FacilityType;
-import com.example.rentalSystem.global.exception.custom.CustomException;
-import com.example.rentalSystem.global.response.ErrorType;
+import com.example.rentalSystem.domain.facility.entity.facility.Facility;
 import java.time.LocalTime;
 import java.util.List;
 
@@ -21,20 +18,18 @@ public record CreateFacilityRequestDto(
 ) {
 
     public Facility toFacility(List<String> imageUrlList) {
-        if (FacilityType.existsByValue(facilityType)) {
-            return Facility.builder()
-                .facilityType(facilityType)
-                .facilityNumber(facilityNumber)
-                .images(imageUrlList)
-                .capacity(capacity)
-                .allowedBoundary(allowedBoundary)
-                .supportFacilities(supportFacilities)
-                .pic(pic)
-                .startTime(startTime)
-                .endTime(endTime)
-                .isAvailable(isAvailable)
-                .build();
-        }
-        throw new CustomException(ErrorType.INVALID_REQUEST);
+        return Facility.builder()
+            .facilityType(facilityType)
+            .facilityNumber(facilityNumber)
+            .images(imageUrlList)
+            .capacity(capacity)
+            .allowedBoundary(allowedBoundary)
+            .supportFacilities(supportFacilities)
+            .pic(pic)
+            .startTime(startTime)
+            .endTime(endTime)
+            .isAvailable(isAvailable)
+            .build();
     }
 }
+
