+++ src/main/java/com/example/rentalSystem/domain/facility/entity/facility/Facility.java
@@ -1,4 +1,4 @@
-package com.example.rentalSystem.domain.facility.entity;
+package com.example.rentalSystem.domain.facility.entity.facility;
 
 import com.example.rentalSystem.domain.common.BaseTimeEntity;
 import com.example.rentalSystem.domain.facility.convert.StringListConverter;
@@ -74,8 +74,7 @@ public class Facility extends BaseTimeEntity {
         LocalTime startTime,
         LocalTime endTime,
         boolean isAvailable) {
-
-        this.facilityType = facilityType;
+        this.facilityType = FacilityType.existsByValue(facilityType);
         this.facilityNumber = facilityNumber;
         this.images = images;
         this.capacity = capacity;
